move most stream components from ContractsFetch to new library
